restructured files so that algos and data structures go into lib, added eslintrc file
