+++ .eslintrc
@@ -1,0 +1,42 @@
+ {
+ "globals": {
+   "describe": false,
+   "it": false,
+   "beforeEach": false,
+   "afterEach": false,
+   "before": false,
+   "after": false
+ },
+  "settings": {
+    "ecmascript": 6
+ },
+ "ecmaFeatures": {
+   "modules": true
+ },
+ "env": {
+   "browser": true,
+   "jquery": true,
+   "node": true
+ },
+ "parserOptions":{
+   "ecmaVersion": 6,
+   "sourceType": "module",
+   "ecmaFeatures": {"jsx": true}
+ },
+ "rules": {
+   "semi": 1,
+   "strict": 0,
+   "indent": [2, 4],
+   "eqeqeq": 1,
+   "quotes": [1, "single"],
+   "no-multi-spaces": [1, {
+     "exceptions": {
+       "VariableDeclarator": true,
+       "FunctionExpression": true
+     }
+   }],
+   "key-spacing": [0, {"align": "value"}],
+   "no-underscore-dangle": 0,
+   "no-unused-vars": 1
+ }
+}
