+++ lib/add.js
@@ -4,12 +4,12 @@ module.exports = {
             return array.reduce((acc,curr) => {
                 return acc + curr;
             }, 0);
-        }
+        };
     },
     sumRecursive: function() {
         return function sumRecursive(array, currIndex = array.length - 1, total = 0) {
             if (!currIndex) return total += array[currIndex];
             else return sumRecursive(array, currIndex - 1, total += array[currIndex]);
-        }
+        };
     },
-}
+};
