+++ lib/binary-max-heap.js
