+++ lib/queue.js
@@ -56,5 +56,5 @@ module.exports = function() {
         checkEmpty() {
             this.checkHead();
         }
-    }
+    };
 };
