+++ lib/random.js
